+++ tools/perf/perf.h
@@ -227,7 +227,7 @@ struct perf_record_opts {
 	unsigned int freq;
 	unsigned int mmap_pages;
 	unsigned int user_freq;
-	u64          branch_stack;
+	int	     branch_stack;
 	u64	     default_interval;
 	u64	     user_interval;
 };
